# docs/openapi.yaml
openapi: 3.1.0
info:
  title: NPLatform API
  version: 1.0.0
  description: |
    국내 최초 AI 기반 NPL(부실채권) 투자 분석 및 거래 플랫폼 API.

    ## 인증
    Supabase JWT를 Authorization 헤더에 Bearer 토큰으로 전달합니다.
    ```
    Authorization: Bearer <supabase-jwt>
    ```

    ## 에러 응답 형식
    모든 에러 응답은 동일한 형식을 따릅니다:
    ```json
    { "error": { "code": "NOT_FOUND", "message": "리소스를 찾을 수 없습니다." } }
    ```

  contact:
    name: NPLatform 개발팀
    email: dev@nplatform.co.kr
  license:
    name: Proprietary
servers:
  - url: https://nplatform.co.kr/api/v1
    description: Production
  - url: http://localhost:3000/api/v1
    description: Local development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - BAD_REQUEST
                - UNAUTHORIZED
                - FORBIDDEN
                - NOT_FOUND
                - CONFLICT
                - RATE_LIMITED
                - VALIDATION_ERROR
                - INTERNAL_ERROR
                - MISSING_FIELD
                - INSUFFICIENT_CREDITS
            message:
              type: string
...
    DealListing:
      type: object
      properties:
        id:
          type: string
          format: uuid
        listing_number:
          type: string
          example: NPL-202601-4382
        title:
          type: string
        listing_type:
          type: string
          enum: [NPL, REO, UPL]
        collateral_type:
          type: string
          enum: [아파트, 오피스텔, 다세대, 단독주택, 상가, 오피스, 토지, 공장, 호텔, 기타]
        address:
          type: string
        principal_amount:
          type: integer
          description: 채권원금 (원)
        appraised_value:
          type: integer
          description: 감정가 (원)
        ltv_ratio:
          type: number
          description: LTV (%)
        risk_grade:
          type: string
          enum: [AAA, AA, A, BBB, BB, B, CCC, CC, C, D]
        status:
          type: string
          enum: [PENDING_REVIEW, ACTIVE, SOLD, EXPIRED, REJECTED]
        visibility:
          type: string
          enum: [PUBLIC, INTERNAL, TARGETED, VIP]
        created_at:
          type: string
          format: date-time
paths:
...
  /exchange/listings/{id}:
    get:
      tags: [listings]
      summary: 매물 상세 조회
      operationId: getListing
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: 매물 상세
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/DealListing'
        '404':
          description: 매물 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

  # ── Statistics ────────────────────────────────────────────
...